+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/layers/interning/ScopedInterningLayer.java
@@ -136,8 +136,7 @@ public class ScopedInterningLayer extends LocatableResolver {
     @Override
     @SuppressWarnings("PMD.CloseResource")
     protected CompletableFuture<Optional<String>> readReverse(FDBStoreTimer timer, Long value) {
-        FDBRecordContext context = database.openContext();
-        context.setTimer(timer);
+        FDBRecordContext context = database.openContext(null, timer);
         return readReverse(context, value)
                 .whenComplete((ignored, th) -> context.close());
     }